+++ FuturaeKit.xcframework/ios-arm64/FuturaeKit.framework/Headers/FTRAccount.h
@@ -30,6 +30,8 @@
 @property (strong, nonatomic) NSArray *allowed_factors;
 @property (strong, nonatomic) NSArray *sessions;
 @property (nonatomic) Boolean logout_pending;
+@property (strong, nonatomic) NSString *hotp_seed;
+@property (nonatomic) NSNumber *sync_counter;
 
 - (instancetype)initWithJSONString:(NSString *)JSONString;
 - (instancetype)initWithDictionary:(NSDictionary*)dictionary;